Move filtering to REST API.
